# services/recipe/src/main/resources/openapi/recipes-v1.openapi.yaml
openapi: "3.0.0"
info:
  title: What's Cookin' API overview
  description: API used to store recipes
  version: 0.1.0
paths:
  /recipes:
    get:
      summary: Returns all recipes from the system that the user has access to
      operationId: findRecipes
      parameters:
        - name: pageSize
          in: query
          description: page size to return (max 100)
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: A paged array of recipes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipes"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create an recipe
      description: Create a recipe desc
      operationId: createRecipe
      tags:
        - recipes
      responses:
        '201':
          description: Null response
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /recipes/{recipeId}:
    get:
      summary: Retreive an employ with ID
      operationId: findRecipeById
      tags:
        - recipes
      parameters:
        - name: recipeId
components:
  schemas:
    PageInfo:
      type: object
      properties:
        number:
          type: integer
          format: int32
          minimum: 0
        total:
          type: integer
          format: int32
          minimum: 0
    Recipe:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    Recipes:
      type: object
      properties:
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
        payload:
          items:
            $ref: '#/components/schemas/Recipe'
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
  parameters:
    pageSizeParam:
      name: pageSize
      in: query
      description: size of page
      schema:
        type: integer
        format: int32
    pageParam:
      name: page
      in: query
      description: page number
      schema:
        type: integer
        format: int32